test(db): open and delete commands
- Cumbersome code to open and delete the AmidzDatabase is implemented as
  the follwong Cypress commands.
    - `openAmidzDatabase`
    - `deleteAmidzDatabase`

  They are defined in a new support file `amidz-database.js`.

issue #60
